Revert "move instrument-selection service to the correct folder"
This reverts commit db45c9d70440deeb877ae1f6424fc8ba64edc1be.
